allow owned strings in request types
Have UserSearchRequest and LevelsRequest store their search strings in a
Cow<'a, str> so that we can put owned strings into there. For example,
when searching for levels by ID, its common to pass in some sort of
`level_id.to_string()` where `level_id` is an u64, but with the current
solution one would need to first store the string on the stack and then
take a reference.

Signed-off-by: stadust <43299462+stadust@users.noreply.github.com>